Implement doxygen documentation.

+++ Source/Utilities/src/RGBColor.cpp
@@ -32,24 +32,24 @@ const RGBColor RGBColor::Red(1.0f, 0.0f, 0.0f);
 const RGBColor RGBColor::Green(0.0f, 1.0f, 0.0f);
 const RGBColor RGBColor::Blue(0.0f, 0.0f, 1.0f);
 
-RGBColor::RGBColor()
-	: r(0.0f),
-	  g(0.0f),
-	  b(0.0f) {
+RGBColor::RGBColor() :
+	r(0.0f),
+	g(0.0f),
+	b(0.0f) {
 
 }
 
-RGBColor::RGBColor(const double new_r, const double new_g, const double new_b)
-	: r(new_r),
-	  g(new_g),
-	  b(new_b) {
+RGBColor::RGBColor(const float new_r, const float new_g, const float new_b) :
+	r(new_r),
+	g(new_g),
+	b(new_b) {
 
 }
 
-RGBColor::RGBColor(const double color)
-	: r(color),
-	  g(color),
-	  b(color) {
+RGBColor::RGBColor(const float color) :
+	r(color),
+	g(color),
+	b(color) {
 
 }
